agent/requirements.txt
boto3>=1.34.0
opencv-python-headless>=4.9.0
numpy>=1.24.0
python-dotenv>=1.0.0
